:construction: Flexible Period Knowledge Section
Rewrited Knowledge Section Knowledge skills' period logic, now it's flexible. It takes date and then filters it to show how many monts/years I have this skill, if the period is less than month it will display - Very short time, also it is possible to display this in case of error date.